references/external_code/airfortomorrow-main/requirements.txt
adjustText==1.3.0
affine==2.4.0
annotated-types==0.7.0
anyio==4.10.0
argcomplete==3.6.2
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
array-api-compat==1.12.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
aws-requests-auth==0.4.3
awswrangler==3.12.1
babel==2.17.0
beautifulsoup4==4.13.4
black==25.1.0
bleach==6.2.0
bokeh==3.7.3
boto3==1.40.11
botocore==1.40.11
branca==0.8.1
Cartopy==0.25.0
cdsapi==0.7.6
certifi==2025.8.3
cffi==1.17.1
cfgrib==0.9.15.0
cftime==1.6.4.post1
charset-normalizer==3.4.3
click==8.2.1
click-plugins==1.1.1.2
cligj==0.7.2
cloudpickle==3.1.1
comm==0.2.3
conflator==0.1.7
contourpy==1.3.2
covjson-pydantic==0.7.0
covjsonkit==0.2.1
cycler==0.12.1
dask==2025.7.0
debugpy==1.8.16
decorator==5.2.1
defusedxml==0.7.1
deprecation==2.1.0
dill==0.4.0
distributed==2025.7.0
docstring_parser==0.17.0
earthkit==0.12.0
earthkit-data==0.16.1
earthkit-geo==0.3.0
earthkit-hydro==0.1.4
earthkit-meteo==0.4.1
earthkit-plots==0.3.3
earthkit-plots-default-styles==0.1.0
earthkit-regrid==0.4.0
earthkit-time==0.1.7
earthkit-transforms==0.4.0
earthkit-utils==0.1.1
earthkit-workflows==0.4.6
eccodes==2.42.0
ecmwf-api-client==1.6.5
ecmwf-datastores-client==0.4.0
ecmwf-opendata==0.3.22
entrypoints==0.4
exactextract==0.2.2
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.116.1
fastcore==1.8.7
fastjsonschema==2.21.2
fastprogress==1.0.3
filelock==3.19.1
findlibs==0.1.2
fiona==1.10.1
fire==0.7.0
flake8==7.3.0
flexcache==0.3
flexparser==0.4
folium==0.20.0
fonttools==4.59.1
fqdn==1.5.1
fsspec==2025.7.0
GDAL==3.6.2
geopandas==1.1.1
geowrangler==0.5.1
h11==0.16.0
h3==3.7.7
h3ronpy==0.21.1
hda==2.34
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
ipykernel==6.30.1
ipython==8.37.0
ipywidgets==8.1.7
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
json5==0.12.1
jsonpickle==4.1.1
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.6
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.9
lark==1.2.2
locket==1.0.0
loguru==0.7.3
lru-dict==1.3.0
lz4==4.4.4
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mistune==3.1.3
morecantile==6.2.0
msgpack==1.1.1
multiurl==0.3.7
mypy_extensions==1.1.0
narwhals==2.1.2
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
netCDF4==1.7.2
networkx==3.4.2
notebook==7.4.5
notebook_shim==0.2.4
numpy==1.26.4
openaq==0.4.0
orjson==3.11.2
overrides==7.7.0
packaging==25.0
pandas==2.3.1
pandocfilters==1.5.1
parso==0.8.4
partd==1.4.2
pathspec==0.12.1
pdbufr==0.14.0
pexpect==4.9.0
pillow==11.3.0
Pint==0.24.4
platformdirs==4.3.8
plotly==6.3.0
pluggy==1.6.0
polars==1.35.1
polars-h3==0.5.7
polytope-client==0.7.7
prometheus_client==0.22.1
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyfdb==0.1.2
pyflakes==3.4.0
Pygments==2.19.2
PyKrige==1.7.2
pyodc==1.5.0
pyogrio==0.11.1
pyparsing==3.2.3
pyproj==3.7.1
pyrsistent==0.20.0
pyshp==2.3.1
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
pytz==2025.2
pyvis==0.3.2
PyYAML==6.0.2
pyzmq==27.0.1
rasterio==1.4.3
rasterstats==0.20.0
referencing==0.36.2
requests==2.32.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.1.0
rich-argparse==1.7.1
rioxarray==0.19.0
rpds-py==0.27.0
s3transfer==0.13.1
scikit-learn==1.7.1
scipy==1.15.3
seaborn==0.13.2
Send2Trash==1.8.3
shapely==2.1.1
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
stack-data==0.6.3
starlette==0.47.2
tblib==3.1.0
termcolor==3.1.0
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tomli==2.2.1
tomlkit==0.13.3
toolz==1.0.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20250809
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.35.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.14
xarray==2025.6.1
xgboost==3.0.4
xmltodict==0.14.2
xyzservices==2025.4.0
yq==3.4.3
zict==3.0.0
zipp==3.23.0
zmq==0.0.0